admin/tag完成，同时完善admin/blogs的添加blog同时，添加blog_tag关系

+++ MyBlog.Model/mapper/Blog_Tag.cs
@@ -10,8 +10,8 @@ namespace MyBlog.Model.mapper
     [SugarTable("Blog_Tag")]
     public class Blog_Tag
     {
-        public int BlogId { get; set; }
+        public long BlogId { get; set; }
 
-        public int TagId { get; set; }
+        public long TagId { get; set; }
     }
 }
